pyproject.toml
[tool.poetry]
name = "keyvox-python-example"
version = "0.1.0"
description = ""
authors = ["keyvox-dev <157120239+keyvox-dev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
keyvox = "^0.1.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
